services/backend/requirements.txt
flask>=3.0.0
pyyaml>=6.0.1
requests>=2.31.0
opencv-python-headless>=4.8.0
pillow>=10.1.0
numpy>=1.26.0
piper-tts>=1.2.0
onnxruntime>=1.16.0
